Suppression de la classe "SudokuSolver"
